move grape publication api to rails controller; remove grape as dependency
