--- lib/sul_bib.rb
@@ -1,2 +1,0 @@
-require 'sul_bib/bib_json_parser'
-require 'sul_bib/year_check'
