--- lib/sul_bib/bib_json_parser.rb
@@ -1,7 +1,0 @@
-module SulBib
-  class BibJSONParser
-    def self.call(object, _env)
-      { pub_hash: JSON.parse(object) }
-    end
-  end
-end
