+++ spec/controllers/authors_controller_spec.rb
@@ -1,4 +1,3 @@
-
 describe AuthorsController do
   let(:author) { create(:author) }
   describe 'POST harvest' do
