+++ spec/models/user_submitted_source_record_spec.rb
@@ -11,7 +11,7 @@ describe UserSubmittedSourceRecord, type: :model do
       expect(user_submitted_source_record).to be_a described_class
     end
     it 'is BibJSON' do
-      json = SulBib::BibJSONParser.call(user_submitted_source_record.source_data, nil)
+      json = { pub_hash: JSON.parse(user_submitted_source_record.source_data) }
       expect(json).to be_a Hash
       expect(json).to include(:pub_hash)
       expect(json[:pub_hash]).to be_a Hash
